essay: creator's block
this is an essay where i write about how i feel so enraged with
myself. Enraged because I wasted decades of my life not creating.
From this moment on, I will devote my life to building. To making
something that people want. I will not, necessarily, create a startup.
There will always be time for that. But, I will not allow my brain
to rot in one of the most beautiful ages for humans--at least
computing-wise.

It was the best of times, it was the worst of times.
Computers were fighting COVID-19. Humanity was fighting itself.
Thankfully, our DNA evolved to learn from scars and wounds.